+++ src/test/java/com/ejoongseok/wmslive/product/feature/Category.java
@@ -1,0 +1,10 @@
+package com.ejoongseok.wmslive.product.feature;
+
+public enum Category {
+    ELECTRONICS("전자 제품");
+    private final String description;
+
+    Category(final String description) {
+        this.description = description;
+    }
+}
