+++ src/test/java/com/ejoongseok/wmslive/product/feature/ProductRepository.java
@@ -1,0 +1,21 @@
+package com.ejoongseok.wmslive.product.feature;
+
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+
+public class ProductRepository {
+    private final Map<Long, Product> products = new HashMap<>();
+    private Long nextId = 1L;
+
+    public void save(final Product product) {
+        product.assignId(nextId);
+        nextId++;
+        products.put(product.getId(), product);
+    }
+
+    public List<Product> findAll() {
+        return new ArrayList<>(products.values());
+    }
+}
